+++ src/main/resources/db/migration/V7__create_table_salesOrder.sql
@@ -1,6 +1,6 @@
 CREATE TABLE Sales_Order(
     id INTEGER AUTO_INCREMENT PRIMARY KEY,
-    price_total DECIMAL(10,2) NOT NULL,
+    price_total Double NOT NULL,
     date_moment TIMESTAMP,
     status INTEGER,
     client_id INTEGER NOT NULL,
